Fixed Filename Order and Added Instructions
The previous versions did not have the correct order for filenames; this is due to its use of the trackFile list during the final writing steps instead of a sorted dictionary. This has been fixed. In addition, I have added instructions on how to include multiple file extensions in the cue sheet.